refactor function
