+++ src/pages/api/user/discoverweekly/archive.ts
@@ -5,26 +5,9 @@ import { withSessionRoute } from "@/lib/withSession";
 import { ArchiveApiResponse } from "@/types/api/user/discoverweekly/archive";
 import { withMongo } from "@/lib/db";
 import { AutoArchiveHistory } from "@/types/db/autoArchiveHistory";
-import { MONGO_DB_COLLECTION_AUTOARCHIVEHISTORY } from "@/lib/common";
+import { getBeginningOfTheWeekDate, MONGO_DB_COLLECTION_AUTOARCHIVEHISTORY } from "@/lib/common";
 
 const playlistIdRegex = /^\w{22}$/
-
-const getDate = (date:Date) => {
-
-  const this_year = date.getFullYear()
-  const this_month = date.getMonth()
-  const this_date = date.getDate()
-  const this_day = date.getDay()
-  const this_monday = this_date - (this_day === 0 ? this_day+6:this_day-1)
-  const WeekStart = new Date(this_year,this_month,this_monday)
-
-  // console.log({this_year,this_month,this_date,this_day,this_monday,WeekStart})
-
-  const fullyear = (d:Date):string => d.getFullYear() + ""
-  const month = (d:Date):string => ((d.getMonth() + 1) < 10 ? "0" : "") + (d.getMonth() + 1) 
-  const day = (d:Date):string => (d.getDate() < 10 ? "0" : "") + d.getDate() 
-  return `${fullyear(WeekStart)}/${month(WeekStart)}/${day(WeekStart)}`
-}
 
 const archive:NextApiHandler<ArchiveApiResponse> = async (req, res) => {
   console.log(`API::${req.method}:${req.url}`,{query:req.query,body:req.body})
@@ -38,7 +21,7 @@ const archive:NextApiHandler<ArchiveApiResponse> = async (req, res) => {
           code:"40301",
           message:"playlistName Error",
         });
-        const weekDate = getDate(new Date())
+        const weekDate = getBeginningOfTheWeekDate(new Date())
         const targetPlaylistName:string = (playlistName as string).includes("{date}") ? (playlistName as string).replace("{date}",weekDate) : playlistName
     
         const targetPlaylistId:string = playlistId
